+++ vocabulo-quizz-app/backend/config/db.js
@@ -1,0 +1,12 @@
+const { Pool } = require('pg');
+
+// Database connection
+const pool = new Pool({
+  user: process.env.POSTGRES_USER,
+  host: process.env.POSTGRES_HOST,
+  database: process.env.POSTGRES_DB,
+  password: process.env.POSTGRES_PASSWORD,
+  port: process.env.POSTGRES_PORT,
+});
+
+module.exports = { pool };
